remove API package from Grakn
