+++ server/src/server/keyspace/KeyspaceImpl.java
@@ -18,20 +18,19 @@
 
 package grakn.core.server.keyspace;
 
-import grakn.core.api.Keyspace;
 import grakn.core.server.exception.TransactionException;
 
 import javax.annotation.CheckReturnValue;
 import java.io.Serializable;
+import java.util.regex.Pattern;
 
-import static grakn.core.api.Keyspace.isValidName;
 
 /**
  * An identifier for an isolated scope of a data in the database.
  */
-public class KeyspaceImpl implements Keyspace, Serializable {
-
+public class KeyspaceImpl implements Serializable, Comparable<KeyspaceImpl> {
     private static final long serialVersionUID = 2726154016735929123L;
+    private static final int MAX_LENGTH = 48;
 
     private final String name;
 
@@ -75,5 +74,15 @@ public class KeyspaceImpl implements Keyspace, Serializable {
         h *= 1000003;
         h ^= this.name.hashCode();
         return h;
+    }
+
+    static boolean isValidName(String name) {
+        return Pattern.matches("[a-z_][a-z_0-9]*", name) && name.length() <= MAX_LENGTH;
+    }
+
+    @Override
+    public int compareTo(KeyspaceImpl o) {
+        if (equals(o)) return 0;
+        return name().compareTo(o.name());
     }
 }
