+++ test-integration/rule/BUILD
@@ -22,7 +22,6 @@ java_library(
     name = "grakn-test-server",
     srcs = ["GraknTestServer.java"],
     deps = [
-        "//api:api",
         "//server",
         "//common",
         "//dependencies/maven/artifacts/junit:junit",
