+++ test-integration/server/kb/concept/BUILD
@@ -90,7 +90,6 @@ java_test(
          "//common:common",
          "//concept:concept",
          "//server:server",
-         "//api:api",
          "@graknlabs_graql//java:graql",
          "//test-integration/rule:grakn-test-server",
          "//dependencies/maven/artifacts/org/hamcrest:hamcrest-library",
